Created skill_fetchers.py, reject suggestion after skill deletion
