+++ core/domain/question_services.py
@@ -24,7 +24,7 @@ from constants import constants
 from core.domain import opportunity_services
 from core.domain import question_domain
 from core.domain import question_fetchers
-from core.domain import skill_services
+from core.domain import skill_fetchers
 from core.domain import state_domain
 from core.platform import models
 import feconf
@@ -362,7 +362,7 @@ def get_skills_linked_to_question(question_id):
         list(Skill). The list of skills that are linked to the question.
     """
     question = get_question_by_id(question_id)
-    skills = skill_services.get_multi_skills(question.linked_skill_ids)
+    skills = skill_fetchers.get_multi_skills(question.linked_skill_ids)
     return skills
