+++ core/domain/skill_services.py
@@ -26,7 +26,9 @@ from core.domain import html_cleaner
 from core.domain import opportunity_services
 from core.domain import role_services
 from core.domain import skill_domain
+from core.domain import skill_fetchers
 from core.domain import state_domain
+from core.domain import suggestion_services
 from core.domain import topic_fetchers
 from core.domain import user_services
 from core.platform import models
@@ -41,117 +43,7 @@ datastore_services = models.Registry.import_datastore_services()
 memcache_services = models.Registry.import_memcache_services()
 
 
-def _migrate_skill_contents_to_latest_schema(versioned_skill_contents):
-    """Holds the responsibility of performing a step-by-step, sequential update
-    of the skill contents structure based on the schema version of the input
-    skill contents dictionary. If the current skill_contents schema changes, a
-    new conversion function must be added and some code appended to this
-    function to account for that new version.
-
-    Args:
-        versioned_skill_contents: A dict with two keys:
-            - schema_version: int. The schema version for the skill_contents
-                dict.
-            - skill_contents: dict. The dict comprising the skill contents.
-
-    Raises:
-        Exception: The schema version of the skill_contents is outside of what
-            is supported at present.
-    """
-    skill_contents_schema_version = versioned_skill_contents['schema_version']
-    if not (1 <= skill_contents_schema_version
-            <= feconf.CURRENT_SKILL_CONTENTS_SCHEMA_VERSION):
-        raise Exception(
-            'Sorry, we can only process v1-v%d skill schemas at '
-            'present.' % feconf.CURRENT_SKILL_CONTENTS_SCHEMA_VERSION)
-
-    while (skill_contents_schema_version <
-           feconf.CURRENT_SKILL_CONTENTS_SCHEMA_VERSION):
-        skill_domain.Skill.update_skill_contents_from_model(
-            versioned_skill_contents, skill_contents_schema_version)
-        skill_contents_schema_version += 1
-
-
-def _migrate_misconceptions_to_latest_schema(versioned_misconceptions):
-    """Holds the responsibility of performing a step-by-step, sequential update
-    of the misconceptions structure based on the schema version of the input
-    misconceptions dictionary. If the current misconceptions schema changes, a
-    new conversion function must be added and some code appended to this
-    function to account for that new version.
-
-    Args:
-        versioned_misconceptions: dict. A dict with two keys:
-            - schema_version: int. The schema version for the misconceptions
-                dict.
-            - misconceptions: list(dict). The list of dicts comprising the skill
-                misconceptions.
-
-    Raises:
-        Exception: The schema version of misconceptions is outside of what
-            is supported at present.
-    """
-    misconception_schema_version = versioned_misconceptions['schema_version']
-    if not (1 <= misconception_schema_version
-            <= feconf.CURRENT_MISCONCEPTIONS_SCHEMA_VERSION):
-        raise Exception(
-            'Sorry, we can only process v1-v%d misconception schemas at '
-            'present.' % feconf.CURRENT_MISCONCEPTIONS_SCHEMA_VERSION)
-
-    while (misconception_schema_version <
-           feconf.CURRENT_MISCONCEPTIONS_SCHEMA_VERSION):
-        skill_domain.Skill.update_misconceptions_from_model(
-            versioned_misconceptions, misconception_schema_version)
-        misconception_schema_version += 1
-
-
-def _migrate_rubrics_to_latest_schema(versioned_rubrics):
-    """Holds the responsibility of performing a step-by-step, sequential update
-    of the rubrics structure based on the schema version of the input
-    rubrics dictionary. If the current rubrics schema changes, a
-    new conversion function must be added and some code appended to this
-    function to account for that new version.
-
-    Args:
-        versioned_rubrics: dict. A dict with two keys:
-            - schema_version: int. The schema version for the rubrics dict.
-            - rubrics: list(dict). The list of dicts comprising the skill
-                rubrics.
-
-    Raises:
-        Exception: The schema version of rubrics is outside of what is supported
-            at present.
-    """
-    rubric_schema_version = versioned_rubrics['schema_version']
-    if not (1 <= rubric_schema_version
-            <= feconf.CURRENT_RUBRIC_SCHEMA_VERSION):
-        raise Exception(
-            'Sorry, we can only process v1-v%d rubric schemas at '
-            'present.' % feconf.CURRENT_RUBRIC_SCHEMA_VERSION)
-
-    while (rubric_schema_version <
-           feconf.CURRENT_RUBRIC_SCHEMA_VERSION):
-        skill_domain.Skill.update_rubrics_from_model(
-            versioned_rubrics, rubric_schema_version)
-        rubric_schema_version += 1
-
-
 # Repository GET methods.
-def _get_skill_memcache_key(skill_id, version=None):
-    """Returns a memcache key for the skill.
-
-    Args:
-        skill_id: str. ID of the skill.
-        version: int or None. Schema version of the skill.
-
-    Returns:
-        str. The memcache key of the skill.
-    """
-    if version:
-        return 'skill-version:%s:%s' % (skill_id, version)
-    else:
-        return 'skill:%s' % skill_id
-
-
 def get_merged_skill_ids():
     """Returns the skill IDs of skills that have been merged.
 
@@ -159,66 +51,6 @@ def get_merged_skill_ids():
         list(str). List of skill IDs of merged skills.
     """
     return [skill.id for skill in skill_models.SkillModel.get_merged_skills()]
-
-
-def get_skill_from_model(skill_model):
-    """Returns a skill domain object given a skill model loaded
-    from the datastore.
-
-    Args:
-        skill_model: SkillModel. The skill model loaded from the datastore.
-
-    Returns:
-        skill. A Skill domain object corresponding to the given skill model.
-    """
-
-    # Ensure the original skill model does not get altered.
-    versioned_skill_contents = {
-        'schema_version': skill_model.skill_contents_schema_version,
-        'skill_contents': copy.deepcopy(skill_model.skill_contents)
-    }
-
-    versioned_misconceptions = {
-        'schema_version': skill_model.misconceptions_schema_version,
-        'misconceptions': copy.deepcopy(skill_model.misconceptions)
-    }
-
-    versioned_rubrics = {
-        'schema_version': skill_model.rubric_schema_version,
-        'rubrics': copy.deepcopy(skill_model.rubrics)
-    }
-
-    # Migrate the skill if it is not using the latest schema version.
-    if (skill_model.skill_contents_schema_version !=
-            feconf.CURRENT_SKILL_CONTENTS_SCHEMA_VERSION):
-        _migrate_skill_contents_to_latest_schema(versioned_skill_contents)
-
-    if (skill_model.misconceptions_schema_version !=
-            feconf.CURRENT_MISCONCEPTIONS_SCHEMA_VERSION):
-        _migrate_misconceptions_to_latest_schema(versioned_misconceptions)
-
-    if (skill_model.rubric_schema_version !=
-            feconf.CURRENT_RUBRIC_SCHEMA_VERSION):
-        _migrate_rubrics_to_latest_schema(versioned_rubrics)
-
-    return skill_domain.Skill(
-        skill_model.id, skill_model.description,
-        [
-            skill_domain.Misconception.from_dict(misconception)
-            for misconception in versioned_misconceptions['misconceptions']
-        ], [
-            skill_domain.Rubric.from_dict(rubric)
-            for rubric in versioned_rubrics['rubrics']
-        ], skill_domain.SkillContents.from_dict(
-            versioned_skill_contents['skill_contents']),
-        versioned_misconceptions['schema_version'],
-        versioned_rubrics['schema_version'],
-        versioned_skill_contents['schema_version'],
-        skill_model.language_code,
-        skill_model.version, skill_model.next_misconception_id,
-        skill_model.superseding_skill_id, skill_model.all_questions_merged,
-        skill_model.prerequisite_skill_ids, skill_model.created_on,
-        skill_model.last_updated)
 
 
 def get_all_skill_summaries():
@@ -488,28 +320,6 @@ def get_multi_skill_summaries(skill_ids):
     return skill_summaries
 
 
-def get_multi_skills(skill_ids, strict=True):
-    """Returns a list of skills matching the skill IDs provided.
-
-    Args:
-        skill_ids: list(str). List of skill IDs to get skills for.
-        strict: bool. Whether to raise an error if a skill doesn't exist.
-
-    Returns:
-        list(Skill). The list of skills matching the provided IDs.
-    """
-    local_skill_models = skill_models.SkillModel.get_multi(skill_ids)
-    for skill_id, skill_model in python_utils.ZIP(
-            skill_ids, local_skill_models):
-        if strict and skill_model is None:
-            raise Exception('No skill exists for ID %s' % skill_id)
-    skills = [
-        get_skill_from_model(skill_model)
-        for skill_model in local_skill_models
-        if skill_model is not None]
-    return skills
-
-
 def get_rubrics_of_skills(skill_ids):
     """Returns a list of rubrics corresponding to given skills.
 
@@ -596,38 +406,6 @@ def get_image_filenames_from_skill(skill):
     """
     html_list = skill.get_all_html_content_strings()
     return html_cleaner.get_image_filenames_from_html_strings(html_list)
-
-
-def get_skill_by_id(skill_id, strict=True, version=None):
-    """Returns a domain object representing a skill.
-
-    Args:
-        skill_id: str. ID of the skill.
-        strict: bool. Whether to fail noisily if no skill with the given
-            id exists in the datastore.
-        version: int or None. The version number of the skill to be
-            retrieved. If it is None, the latest version will be retrieved.
-
-    Returns:
-        Skill or None. The domain object representing a skill with the
-        given id, or None if it does not exist.
-    """
-    skill_memcache_key = _get_skill_memcache_key(
-        skill_id, version=version)
-    memcached_skill = memcache_services.get_multi(
-        [skill_memcache_key]).get(skill_memcache_key)
-
-    if memcached_skill is not None:
-        return memcached_skill
-    else:
-        skill_model = skill_models.SkillModel.get(
-            skill_id, strict=strict, version=version)
-        if skill_model:
-            skill = get_skill_from_model(skill_model)
-            memcache_services.set_multi({skill_memcache_key: skill})
-            return skill
-        else:
-            return None
 
 
 def get_skill_summary_by_id(skill_id, strict=True):
@@ -728,7 +506,7 @@ def apply_change_list(skill_id, change_list, committer_id):
     Returns:
         Skill. The resulting skill domain object.
     """
-    skill = get_skill_by_id(skill_id)
+    skill = skill_fetchers.get_skill_by_id(skill_id)
     user = user_services.UserActionsInfo(committer_id)
     try:
         for change in change_list:
@@ -878,7 +656,7 @@ def _save_skill(committer_id, skill, commit_message, change_list):
     skill_model.next_misconception_id = skill.next_misconception_id
     change_dicts = [change.to_dict() for change in change_list]
     skill_model.commit(committer_id, commit_message, change_dicts)
-    memcache_services.delete(_get_skill_memcache_key(skill.id))
+    memcache_services.delete(skill_fetchers.get_skill_memcache_key(skill.id))
     skill.version += 1
 
 
@@ -926,13 +704,15 @@ def delete_skill(committer_id, skill_id, force_deletion=False):
 
     # This must come after the skill is retrieved. Otherwise the memcache
     # key will be reinstated.
-    skill_memcache_key = _get_skill_memcache_key(skill_id)
+    skill_memcache_key = skill_fetchers.get_skill_memcache_key(skill_id)
     memcache_services.delete(skill_memcache_key)
 
     # Delete the summary of the skill (regardless of whether
     # force_deletion is True or not).
     delete_skill_summary(skill_id)
     opportunity_services.delete_skill_opportunity(skill_id)
+    suggestion_services.reject_question_suggestions_with_skill_target_id(
+        skill_id)
 
 
 def delete_skill_summary(skill_id):
@@ -976,7 +756,7 @@ def create_skill_summary(skill_id):
     Args:
         skill_id: str. ID of the skill.
     """
-    skill = get_skill_by_id(skill_id)
+    skill = skill_fetchers.get_skill_by_id(skill_id)
     skill_summary = compute_summary_of_skill(skill)
     save_skill_summary(skill_summary)
