+++ core/domain/suggestion_registry.py
@@ -26,7 +26,7 @@ from core.domain import exp_services
 from core.domain import question_domain
 from core.domain import question_services
 from core.domain import skill_domain
-from core.domain import skill_services
+from core.domain import skill_fetchers
 from core.domain import state_domain
 from core.domain import user_id_migration
 from core.domain import user_services
@@ -659,7 +659,7 @@ class SuggestionAddQuestion(BaseSuggestion):
                 'Question state schema version is not up to date.')
 
         skill_domain.Skill.require_valid_skill_id(self.change.skill_id)
-        skill = skill_services.get_skill_by_id(
+        skill = skill_fetchers.get_skill_by_id(
             self.change.skill_id, strict=False)
         if skill is None:
             raise utils.ValidationError(
@@ -684,7 +684,7 @@ class SuggestionAddQuestion(BaseSuggestion):
         question = question_domain.Question.from_dict(question_dict)
         question.validate()
         question_services.add_question(self.author_id, question)
-        skill = skill_services.get_skill_by_id(
+        skill = skill_fetchers.get_skill_by_id(
             self.change.skill_id, strict=False)
         if skill is None:
             raise utils.ValidationError(
